Make object_id's persistent across undo/redos, and more
The for_each_point_object* functions have been renamed to for_each_object*
and their implementation is now defined in terms of a new macro called
for_each_object_type which iterates over each type of object. It uses
template lambas to pass the type information back to the 'caller' which
requires C++20.

A bunch of other object system-related code has also been rewritten to
work with the new system. For example, the object picking code has been
modified so that the entire key is stored as the colour of the object
in the pickframe (see view_3d.cpp).