Add stream viewer
This window displays the tree of streams that are currently
in use by Wrench, including a human-friendly name, their
resource path and their size. This can be used to determine
the location of assets on disc. After a few more commits, it
should be able to do a whole lot more too!