models: Call glDrawArrays with the correct index count
Previously the code assumed that moby_vertex_data_to_opengl
would output the same number of vertices as there were elements
in the index buffer, which is false for now since some indices
aren't read in correctly.

This meant that the value passed to glDrawArrays was too
large, creating strange artifacts on screen.