gui: Made dialog boxes more ImGui-like
String input and alert boxes now work a lot more ImGui-like.
For example now to create a text prompt you can do:

	static prompt_box box("Button Text", "Enter String");
	if(auto str = box.prompt()) {
		printf("String entered: %s\n", str->c_str());
	}

This will create a button with the text "Button Text" which
when pressed will pop up a string input box. When the user
presses the "Okay" button in said dialog box, prompt_box::prompt
will return the string entered wrapped in an optional, and the
printf will be run.

I've modified a bunch of code relating to e.g. saving projects
so these new functions are used.