compression: Handle emitting literal packets much better
This includes EOF packets which are now emitted with reasonable
lengths. Normal literal packets will also have a dummy packet
inserted behind them if a small literal size needs to be injected.

I spent hours trying to find a bug that was causing the latter
part not to work. Turned out there was a <= where a < should've been.