Improve error handling in level::read_loading_screen_textures
This gets R&C3 level loading working again. This is more of a
temporary solution, the underlying problem is that the file_stream
object implodes whenever there's the slightest error and can't
recover. I plan on overhauling the patching system at some point
so I think it'll be best to wait until then to properly fix this.