compressor: Remove special case for emitting special literals at the end of a buffer
Previously at the end of an intermediate buffer (for a single thread), only
literals would be emitted so that the match finding code did not have to check
the end condition. This is no longer the case.

For speed, the match finding code is now a function template taking a boolean
template parameter for if it needs to check for the end of the buffer or not.

This increases performance a bit. It also increases compression ratios a
bit although it's only noticeable for smaller files where the number of threads
is high since it reduces a constant-ish overhead that's in terms of the number
of threads.