Improve error handling when importing a level
Previously, if importing a level failed the original level would
still be cleared. To fix this, I've cleaned up how textures are
read in so that levels can be moved while being created. Previously
textures contained a pointer to a stream, which would be
invalidated if the level was moved in memory. Now textures contain
the data they need directly and don't access a stream after they
are created.