iso: When extracting level parts from R&C3/DL, name them correctly
Previously the extractor assumed that the level table was set out
like is is in R&C2. Code has been added to identify the type of a
file based on the magic identifier in its header. Additionally,
the table of contents parsing code has been refactored.

Also also, we now handle files with headers that are larger than
one sector better.