Refactor ISO building code
This cleans up the code a whole bunch. Previously, it would
calculate all the LBAs, write out the filesystem, the write out
the actual data. Now, it calculates the LBAs at the same time as it
writes out the data, then it writes out the filesystem and table of
contents at the end.

The printout has also been redesigned: Now it's a table. The table
of contents file is now also referenced by the filesystem.