wad: Perfect the moby model recovery code
The elements of the duplicate vertices array for each moby submesh
are bitshifted indices into an intermediate buffer. For each vertex
in a model, the intermediate buffer is populated at index equal to
the first 9 bits of the vertex 7 vertices before the current vertex
with the current vertex.

Additionally, the moby VU1 microprogram stores additional
indices in the index header and the texture unpacks. The stopping
condition for the microprogram is when one of these indices is zero.