Fixed some issues related to building with Ninja.
Ninja requires that byproducts like in the case of zlib.lib are
specified. Also the examples of cxxopts are build in a weird way that is
not supported by Ninja so since we don't need those examples I turned
them off. Finally, the DEPENDS "*" thing only works in make. I am not
quite sure what the best way to fix this is so I propose to use this
file(GLOB ... CONFIGURE_DEPENDS ...) approach which is probably under
the hood giving us the same behaviour.