packer: Fixed a bug where files where packed with padded reported dimensions without actual padding.
This was mostly an issue when packing gui.wad as the last packed file was not
padded towards the end. While it is not necessary to pad the end of the
file. However, the reported size of the packed file included padding which caused
file reads beyond the end of file.