Merge pvar types from multiple levels
This prevents false negatives e.g. if for the last level unpacked
that contains a given class, and all instances of that type have
a link field set to -1, that field would previously not have its
type information recovered, even if another level contains evidence
that the field is a link.